refactor(sio): break circular dependency in source code
Related: https://github.com/socketio/socket.io/issues/4329
